+++ test/browser/specs/mocking.test.ts
@@ -1,6 +1,5 @@
 import { expect, test } from 'vitest'
 import { runVitest } from '../../test-utils'
-import { runBrowserTests } from './utils'
 
 test.each([true, false])('mocking works correctly - isolated %s', async (isolate) => {
   const result = await runVitest({
@@ -16,16 +15,6 @@ test.each([true, false])('mocking works correctly - isolated %s', async (isolate
   expect(result.stdout).toContain('mocked-nested.test.ts')
   expect(result.stdout).toContain('not-mocked-nested.test.ts')
   expect(result.stdout).toContain('import-actual-in-mock.test.ts')
+  expect(result.stdout).toContain('import-actual-query.test.ts')
   expect(result.exitCode).toBe(0)
-})
-
-test('mocking fails if fileParallelism is enabled', async () => {
-  const result = await runBrowserTests({
-    root: 'fixtures/mocking',
-    browser: {
-      fileParallelism: true,
-    },
-  })
-  expect(result.stderr).toContain('Mocking doesn\'t work with "browser.fileParallelism" enabled')
-  expect(result.exitCode).toBe(1)
 })
